Add files via upload
I was able to complete most of the work on the Learn to Code webpage, but have struggled with getting items in my Header to align properly.  I know I need to add Mobile Responsiveness, but after talking with Ian Washington the other day and getting an extension I feel I need to get something submitted now.  Any feedback to help me with those items would be greatly appreciated.  I feel like I have a decent understanding of the concepts related to Class and ID Selectors, Inherited Properties, Positioning, Floats, Flex-Direction, and Justify-Content, but am having a hard time putting them into practice, which is stuff I was hoping to learn in the Bootcamp.