Add files via upload
Corrected the mobile-responsiveness as these were flipped previously.  That changes how the Nav List appear based on screen size and whether the Image appears.
I also corrected the way I setup the UL, entering a single UL opening and closing tag, then the Nav List items are entered with multiple LI tags.

The separation of the Nav List items isn't quite where I want it to be.  And I still couldn't get the Buttons to response based on screen size.  But hope with the changes made I can get over the 80% required.